+++ application/server/src/main/kotlin/io/github/siyual_park/application/service/handler/VilificationExceptionHandler.kt
@@ -1,7 +1,7 @@
 package io.github.siyual_park.application.service.handler
 
 import io.github.siyual_park.application.service.exception.BadRequestException
-import io.github.siyual_park.search.exception.InvalidRequestException
+import io.github.siyual_park.reader.exception.InvalidRequestException
 import org.springframework.core.Ordered
 import org.springframework.core.annotation.Order
 import org.springframework.web.bind.annotation.ExceptionHandler
