+++ domain/client/src/main/kotlin/io/github/siyual_park/client/domain/ClientScopeFinder.kt
@@ -4,7 +4,7 @@ import io.github.siyual_park.auth.domain.scope_token.ScopeTokenFinder
 import io.github.siyual_park.auth.entity.ScopeToken
 import io.github.siyual_park.client.entity.Client
 import io.github.siyual_park.client.repository.ClientScopeRepository
-import io.github.siyual_park.search.finder.Finder
+import io.github.siyual_park.reader.finder.Finder
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emitAll
 import kotlinx.coroutines.flow.emptyFlow
