+++ domain/client/src/main/kotlin/io/github/siyual_park/client/domain/SyncClientScope.kt
@@ -10,9 +10,9 @@ import io.github.siyual_park.data.event.AfterSaveEvent
 import io.github.siyual_park.data.expansion.where
 import io.github.siyual_park.event.EventConsumer
 import io.github.siyual_park.event.Subscribe
-import io.github.siyual_park.search.finder.findByIdOrFail
-import io.github.siyual_park.search.pagination.OffsetPaginatorAdapter
-import io.github.siyual_park.search.pagination.forEach
+import io.github.siyual_park.reader.finder.findByIdOrFail
+import io.github.siyual_park.reader.pagination.OffsetPaginatorAdapter
+import io.github.siyual_park.reader.pagination.forEach
 import org.springframework.stereotype.Component
 
 @Component
