+++ module/reader/build.gradle.kts
