+++ module/reader/src/main/kotlin/io/github/siyual_park/reader/exception/SortInvalidException.kt
@@ -1,3 +1,3 @@
-package io.github.siyual_park.search.exception
+package io.github.siyual_park.reader.exception
 
 class SortInvalidException(message: String? = null) : InvalidRequestException(message)
