+++ module/reader/src/main/kotlin/io/github/siyual_park/reader/pagination/OffsetPage.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.search.pagination
+package io.github.siyual_park.reader.pagination
 
 class OffsetPage<T>(
     val data: Collection<T>,
